requirements.txt
./local_wheels/dash_uploader-0.5.0-py3-none-any.whl
alembic
bs4
colorlover
dash
dash-bootstrap-components
dash-extensions
dash-renderer
dash-tabulator
email_validator
filelock
flask
flask-compress
flask-login
flask-login
flask-migrate
flask-wtf
ipyfilechooser
ipywidgets
jsbeautifier
lxml
matplotlib
molmass
numpy
openpyxl
pandas
pandoc
plotly
plotly_express
pyarrow>=3
pymzml
pyteomics
pytest-cov
scikit-learn
scipy==1.6.2
seaborn
setuptools
six
statsmodels
tables
tqdm
urllib3
waitress
wget
xlrd
xlsxwriter
black
nbval